style(components): move `Accordion` to library
